# dashboard/openapi.json
{
  "components": {
    "schemas": {
...
      "AddDeviceIn": {
        "description": "Password is `SecretStr` so an accidental `repr()` anywhere, including\na structlog kwarg, renders `**********` instead of the plaintext.",
        "properties": {
          "alias": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Alias",
            "type": "string"
          },
          "channel": {
            "anyOf": [
              {
                "maximum": 255.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "host": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Host",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "password": {
            "format": "password",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          },
          "timezone": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Timezone",
            "type": "string"
          },
          "user": {
            "maxLength": 100,
            "minLength": 1,
            "title": "User",
            "type": "string"
          }
        },
        "required": [
          "alias",
          "host",
          "user",
          "password",
          "timezone"
        ],
        "title": "AddDeviceIn",
        "type": "object"
      },
...
      "DeviceOut": {
        "properties": {
          "alias": {
            "title": "Alias",
            "type": "string"
          },
          "archived_bytes": {
            "title": "Archived Bytes",
            "type": "integer"
          },
          "archived_count": {
            "title": "Archived Count",
            "type": "integer"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "gateway_down": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gateway Down"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sample": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StorageSampleOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "sd": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SdBarOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "alias",
          "host",
          "model",
          "name",
          "timezone",
          "enabled",
          "archived_count",
          "archived_bytes",
          "sample",
          "sd",
          "last_run",
          "gateway_down"
        ],
        "title": "DeviceOut",
        "type": "object"
      },
...
      "DevicesOut": {
        "properties": {
          "devices": {
            "items": {
              "$ref": "#/components/schemas/DeviceOut"
            },
            "title": "Devices",
            "type": "array"
          },
          "registry_available": {
            "title": "Registry Available",
            "type": "boolean"
          }
        },
        "required": [
          "devices",
          "registry_available"
        ],
        "title": "DevicesOut",
        "type": "object"
      },
      "DiscoveredOut": {
        "properties": {
          "already_registered": {
            "title": "Already Registered",
            "type": "boolean"
          },
          "host": {
            "title": "Host",
            "type": "string"
          },
          "mac": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mac"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "protocol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol"
          },
          "uid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uid"
          }
        },
        "required": [
          "host",
          "uid",
          "mac",
          "protocol",
          "model",
          "name",
          "already_registered"
        ],
        "title": "DiscoveredOut",
        "type": "object"
      },
      "EnabledIn": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "EnabledIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunOut": {
        "properties": {
          "bytes_archived": {
            "title": "Bytes Archived",
            "type": "integer"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "discovered": {
            "title": "Discovered",
            "type": "integer"
          },
          "downloaded": {
            "title": "Downloaded",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "skipped_dup": {
            "title": "Skipped Dup",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          },
          "window_from_utc": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window From Utc"
          },
          "window_to_utc": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window To Utc"
          }
        },
        "required": [
          "id",
          "device_id",
          "trigger",
          "window_from_utc",
          "window_to_utc",
          "started_at",
          "finished_at",
          "outcome",
          "discovered",
          "downloaded",
          "skipped_dup",
          "failed",
          "bytes_archived",
          "error"
        ],
        "title": "RunOut",
        "type": "object"
      },
...
      "SdBarOut": {
        "description": "GiB, all on the camera's own binary scale (see `sd_bar`).",
        "properties": {
          "archived_gb": {
            "title": "Archived Gb",
            "type": "number"
          },
          "free_gb": {
            "title": "Free Gb",
            "type": "number"
          },
          "other_gb": {
            "title": "Other Gb",
            "type": "number"
          },
          "total_gb": {
            "title": "Total Gb",
            "type": "number"
          },
          "used_gb": {
            "title": "Used Gb",
            "type": "number"
          }
        },
        "required": [
          "total_gb",
          "used_gb",
          "free_gb",
          "archived_gb",
          "other_gb"
        ],
        "title": "SdBarOut",
        "type": "object"
      },
...
      "StorageSampleOut": {
        "properties": {
          "formatted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formatted"
          },
          "mounted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mounted"
          },
          "remain_gb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remain Gb"
          },
          "sampled_at": {
            "format": "date-time",
            "title": "Sampled At",
            "type": "string"
          },
          "total_gb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Gb"
          }
        },
        "required": [
          "total_gb",
          "remain_gb",
          "formatted",
          "mounted",
          "sampled_at"
        ],
        "title": "StorageSampleOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ReoVault",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/devices": {
      "get": {
        "operationId": "list_devices_api_v1_devices_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevicesOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Devices",
        "tags": [
          "devices"
        ]
      },
      "post": {
        "operationId": "add_device_api_v1_devices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddDeviceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevicesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Content"
          },
          "502": {
            "description": "Bad Gateway"
          },
          "503": {
            "description": "Service Unavailable"
          }
        },
        "summary": "Add Device",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/discover": {
      "post": {
        "operationId": "discover_api_v1_devices_discover_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DiscoveredOut"
                  },
                  "title": "Response Discover Api V1 Devices Discover Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "502": {
            "description": "Bad Gateway"
          },
          "503": {
            "description": "Service Unavailable"
          }
        },
        "summary": "Discover",
        "tags": [
          "devices"
        ]
      }
    },
...
    "/api/v1/devices/{device_id}/enabled": {
      "put": {
        "operationId": "set_enabled_api_v1_devices__device_id__enabled_put",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnabledIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevicesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Enabled",
        "tags": [
          "devices"
        ]
      }
    },
...
  }
}